feat: Add endpoint to get overview data
